requirements.txt
requests~=2.27.0
Pillow~=9.0.0
PySide6~=6.2.2.1
python-twitter~=3.5